Workaround: Do not report required input field labes as a problem when not properly setup
The required html should be set, but this test tends to produce many false positives.
These many false positives are usually the result of conditionally required fields.

The conditionally required fields are required and on the form, but are only displayed depending on certain situations.

For the time being, suppress this area and refer to it as a suggestion until a proper solution may be implemented.