Update: Add the html class 'wysiwyg wysiwyg-ckeditor' to the embeded html
This is required to get css styling specific to the ckeditor wysiwyg area to properly function.
As far as I can tell, there was no provided means to do this other than editing the source of ckeditor.
Therefore this step has to be re-performed when this module gets updated.

Here is how to do what I did:
1) modify ckeditor/_source/plugins/wysiwygarea/plugin.js:
- Find all cases of "<html>" and "<html " and add
  class="wysiwyg wysiwyg-ckeditor".
  (Such that you get "<html class="wysiwyg wysiwyg-ckeditor">"
2) re-package/compile the ckeditor _source into the ckeditor/ckeditor.js:
  a) Get the CKPackager from the ckeditor website:
    - http://docs.cksource.com/CKEditor_3.x/Developers_Guide/CKPackager
    - http://svn.ckeditor.com/CKPackager/trunk/bin/ckpackager.jar
  b) From inside the ckeditor directory, execute the packager:
    - java -jar ckpackager.jar ckeditor.pack

This is necessary to help solve issue www-26, but it also provides us with a means to make ckeditor more precicely match what the content as it is shown (while editing).

ticket: www-26